requirements.txt
PySide6
pyqtgraph
pandas
numpy
scipy
lmfit
matplotlib
pydantic
pytest-qt
libgl1
